fix: 修复订单ID类型不匹配问题
- 将IOrderService接口中所有orderId参数从long改为int
- 将OrderService实现中所有orderId参数从long改为int
- 将OrdersController中所有id参数从long改为int

修复原因:
- Order.Id属性类型为int
- OrderReview.OrderId属性类型为int
- 服务方法和控制器参数使用了long类型,导致类型不匹配
- CS0266编译错误: 无法将long隐式转换为int

修改的文件:
- Services/IOrderService.cs - 5处方法签名
- Services/OrderService.cs - 所有方法实现
- Controllers/OrdersController.cs - 5处方法签名

现在所有订单ID参数类型与实体属性类型一致,编译错误已修复。

+++ Controllers/OrdersController.cs
@@ -80,7 +80,7 @@ public class OrdersController : ControllerBase
     [ProducesResponseType(typeof(ApiResponse<GetOrderResponse>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status403Forbidden)]
-    public async Task<ActionResult<ApiResponse<GetOrderResponse>>> GetOrder(long id)
+    public async Task<ActionResult<ApiResponse<GetOrderResponse>>> GetOrder(int id)
     {
         _logger.LogInformation("用户获取订单详情: OrderId={OrderId}", id);
 
@@ -98,7 +98,7 @@ public class OrdersController : ControllerBase
     [ProducesResponseType(typeof(ApiResponse<CancelOrderResponse>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status422UnprocessableEntity)]
-    public async Task<ActionResult<ApiResponse<CancelOrderResponse>>> CancelOrder(long id, [FromBody] CancelOrderRequest request)
+    public async Task<ActionResult<ApiResponse<CancelOrderResponse>>> CancelOrder(int id, [FromBody] CancelOrderRequest request)
     {
         _logger.LogInformation("用户取消订单: OrderId={OrderId}, Reason={Reason}", id, request.CancelReason);
 
@@ -116,7 +116,7 @@ public class OrdersController : ControllerBase
     [ProducesResponseType(typeof(ApiResponse<RefundOrderResponse>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status422UnprocessableEntity)]
-    public async Task<ActionResult<ApiResponse<RefundOrderResponse>>> RefundOrder(long id, [FromBody] RefundOrderRequest request)
+    public async Task<ActionResult<ApiResponse<RefundOrderResponse>>> RefundOrder(int id, [FromBody] RefundOrderRequest request)
     {
         _logger.LogInformation("用户申请退款: OrderId={OrderId}, Amount={Amount}", id, request.RefundAmount);
 
@@ -133,7 +133,7 @@ public class OrdersController : ControllerBase
     [ProducesResponseType(typeof(ApiResponse<ConfirmOrderResponse>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status422UnprocessableEntity)]
-    public async Task<ActionResult<ApiResponse<ConfirmOrderResponse>>> ConfirmOrder(long id)
+    public async Task<ActionResult<ApiResponse<ConfirmOrderResponse>>> ConfirmOrder(int id)
     {
         _logger.LogInformation("用户确认订单完成: OrderId={OrderId}", id);
 
@@ -151,7 +151,7 @@ public class OrdersController : ControllerBase
     [ProducesResponseType(typeof(ApiResponse<CreateOrderReviewResponse>), StatusCodes.Status200OK)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status404NotFound)]
     [ProducesResponseType(typeof(ApiResponse), StatusCodes.Status422UnprocessableEntity)]
-    public async Task<ActionResult<ApiResponse<CreateOrderReviewResponse>>> ReviewOrder(long id, [FromBody] CreateOrderReviewRequest request)
+    public async Task<ActionResult<ApiResponse<CreateOrderReviewResponse>>> ReviewOrder(int id, [FromBody] CreateOrderReviewRequest request)
     {
         _logger.LogInformation("用户订单评价: OrderId={OrderId}, Rating={Rating}", id, request.Rating);
